+++ src/test/java/cart/controller/admin/AdminControllerTest.java
@@ -1,0 +1,54 @@
+package cart.controller.admin;
+
+import cart.controller.dto.ProductDto;
+import cart.persistence.entity.ProductCategory;
+import cart.service.ShoppingService;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
+import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.test.web.servlet.MockMvc;
+
+import java.util.List;
+
+import static org.mockito.Mockito.when;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+
+@WebMvcTest(AdminController.class)
+class AdminControllerTest {
+
+    private ProductDto productDto;
+
+    @Autowired
+    private MockMvc mockMvc;
+
+    @Autowired
+    private ObjectMapper objectMapper;
+
+    @MockBean
+    private ShoppingService shoppingService;
+
+    @BeforeEach
+    void setUp() {
+        productDto = new ProductDto(1L, "치킨", "chickenUrl", 20000, ProductCategory.KOREAN);
+    }
+
+    @Test
+    void getProducts() throws Exception {
+        // given
+        final List<ProductDto> productDtos = List.of(
+                new ProductDto(1L, "치킨", "chickenUrl", 20000, ProductCategory.KOREAN),
+                new ProductDto(2L, "초밥", "chobobUrl", 30000, ProductCategory.JAPANESE),
+                new ProductDto(3L, "스테이크", "steakUrl", 40000, ProductCategory.WESTERN)
+        );
+        when(shoppingService.getProducts()).thenReturn(productDtos);
+
+        // when, then
+        mockMvc.perform(get("/admin"))
+                .andExpect(status().isOk());
+    }
+
+}
